+++ systems/setonix/custom/quantum_install_scripts/install-pennylane-source-setonix.sh
@@ -2,17 +2,32 @@
 
 script_dir="$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 . $script_dir/use-pennylane-source-setonix.sh
-cd ${base_dir}
 
 # install from source
-git clone https://github.com/PennyLaneAI/pennylane $source_dir
-cd $source_dir
-git fetch --tags 
-git checkout v$tool_ver
 
-CMAKE_ARGS="-DCMAKE_CXX_COMPILER=CC \
-  -DCMAKE_BUILD_TYPE=Release \
-  pip install --prefix=$install_dir .
+# if no argument is passed then assuming a build is required
+if [ -z $1 ]; then 
+    echo "Building ${tool_name}/${tool_ver}"
+    # install from source
+    # note that this repo is deprecated and now should use pennylane-lightning
+    # but I cannot get that version to compile
+    cd ${base_dir}
+    git clone https://github.com/PennyLaneAI/pennylane $source_dir
+    cd $source_dir
+    git fetch --tags 
+    git checkout v$tool_ver
 
-cd -
-rm -rf ${source_dir}
+    CMAKE_ARGS="-DCMAKE_CXX_COMPILER=CC \
+        -DCMAKE_BUILD_TYPE=Release"
+    cmake -B build . ${CMAKE_ARGS}
+    cmake --build build
+    pip install --prefix=$install_dir .
+
+    cd -
+    rm -rf ${source_dir}
+fi
+
+# install module 
+${script_dir}/install-module.sh ${install_dir} \
+${tool_name} ${tool_ver} \
+"${brief}" "${descrip}"
